Extract pre-model preparation concurrency
