+++ src/mindroom/ai.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import asyncio
-import contextvars
 from collections.abc import Mapping
 from contextlib import aclosing
 from dataclasses import dataclass, field
@@ -68,6 +67,7 @@ from mindroom.media_fallback import (
 from mindroom.media_inputs import MediaInputs, MediaKind
 from mindroom.memory import MemoryPromptParts, build_memory_prompt_parts, strip_user_turn_time_prefix
 from mindroom.metadata_merge import deep_merge_metadata
+from mindroom.pre_model_preparation import prepare_mem0_prompt_branches
 from mindroom.response_turn import (
     AttemptResolved,
     BlockingAttemptResolution,
@@ -88,7 +88,7 @@ from mindroom.timing import DispatchPipelineTiming, emit_timing_event, timed, ti
 from mindroom.tool_system.events import StreamingToolTracker, complete_pending_tool_block, format_tool_combined
 
 if TYPE_CHECKING:
-    from collections.abc import AsyncGenerator, AsyncIterator, Awaitable, Callable, Sequence
+    from collections.abc import AsyncGenerator, AsyncIterator, Callable, Sequence
     from contextlib import AbstractContextManager
 
     from agno.agent import Agent
@@ -1066,134 +1066,6 @@ def _mark_pipeline_timing(pipeline_timing: DispatchPipelineTiming | None, label:
         pipeline_timing.mark(label)
 
 
-def _close_unreturned_agent(agent: Agent, scope_context: ScopeSessionContext | None) -> None:
-    """Close runtime state owned by an agent that preparation cannot return."""
-    try:
-        close_agent_runtime_state_dbs(
-            agent,
-            shared_scope_storage=scope_context.storage if scope_context is not None else None,
-        )
-    except Exception:
-        logger.exception("Failed to close unreturned agent runtime state", agent=agent.id)
-
-
-async def _drain_unreturned_agent_build(
-    build_future: asyncio.Future[tuple[ResolvedRuntimeModel, Agent]],
-    *,
-    cleanup_agent: Callable[[Agent], None],
-    on_error: Callable[[Exception], None],
-) -> None:
-    """Wait through repeated cancellation and clean an unreturned agent build."""
-    while not build_future.done():
-        try:
-            await asyncio.shield(build_future)
-        except asyncio.CancelledError:
-            continue
-        except Exception:
-            break
-    try:
-        _, unreturned_agent = build_future.result()
-    except asyncio.CancelledError:
-        return
-    except Exception as error:
-        on_error(error)
-    else:
-        cleanup_agent(unreturned_agent)
-
-
-def _discard_unreturned_agent_result(
-    result: tuple[ResolvedRuntimeModel, Agent] | Exception,
-    *,
-    cleanup_agent: Callable[[Agent], None],
-    on_error: Callable[[Exception], None],
-) -> None:
-    """Log a failed build or close an agent that preparation cannot return."""
-    if isinstance(result, Exception):
-        on_error(result)
-    else:
-        cleanup_agent(result[1])
-
-
-async def _prepare_mem0_branches(
-    *,
-    prepare_memory: Callable[[], Awaitable[MemoryPromptParts]],
-    build_agent: Callable[[], tuple[ResolvedRuntimeModel, Agent]],
-    cleanup_agent: Callable[[Agent], None],
-    agent_name: str,
-    pipeline_timing: DispatchPipelineTiming | None,
-) -> tuple[MemoryPromptParts, ResolvedRuntimeModel, Agent]:
-    """Overlap Mem0 preparation with agent construction and join both safely."""
-
-    def _log_secondary_agent_error(error: Exception) -> None:
-        logger.error(
-            "Agent construction failed while memory preparation was unavailable",
-            agent=agent_name,
-            error=repr(error),
-        )
-
-    async def _memory_branch() -> MemoryPromptParts | Exception:
-        _mark_pipeline_timing(pipeline_timing, "memory_prepare_start")
-        try:
-            return await prepare_memory()
-        except Exception as error:
-            return error
-        finally:
-            _mark_pipeline_timing(pipeline_timing, "memory_prepare_ready")
-
-    context = contextvars.copy_context()
-    _mark_pipeline_timing(pipeline_timing, "agent_build_start")
-    build_future = asyncio.get_running_loop().run_in_executor(None, context.run, build_agent)
-    build_future.add_done_callback(
-        lambda _future: _mark_pipeline_timing(pipeline_timing, "agent_build_ready"),
-    )
-
-    async def _agent_branch() -> tuple[ResolvedRuntimeModel, Agent] | Exception:
-        try:
-            return await asyncio.shield(build_future)
-        except Exception as error:
-            return error
-
-    try:
-        async with asyncio.TaskGroup() as task_group:
-            agent_task = task_group.create_task(
-                _agent_branch(),
-                name=f"agent_prepare:{agent_name}",
-            )
-            memory_task = task_group.create_task(
-                _memory_branch(),
-                name=f"memory_prepare:{agent_name}",
-            )
-    except BaseException:
-        await _drain_unreturned_agent_build(
-            build_future,
-            cleanup_agent=cleanup_agent,
-            on_error=_log_secondary_agent_error,
-        )
-        raise
-
-    agent_result = agent_task.result()
-    try:
-        memory_result = memory_task.result()
-    except BaseException:
-        _discard_unreturned_agent_result(
-            agent_result,
-            cleanup_agent=cleanup_agent,
-            on_error=_log_secondary_agent_error,
-        )
-        raise
-    if isinstance(memory_result, Exception):
-        _discard_unreturned_agent_result(
-            agent_result,
-            cleanup_agent=cleanup_agent,
-            on_error=_log_secondary_agent_error,
-        )
-        raise memory_result
-    if isinstance(agent_result, Exception):
-        raise agent_result
-    runtime_model, agent = agent_result
-    return memory_result, runtime_model, agent
-
-
 @timed("system_prompt_assembly")
 async def _prepare_agent_and_prompt(
     ctx: ResponseTurnContext,
@@ -1262,7 +1134,7 @@ async def _prepare_agent_and_prompt(
             parallel=True,
         ):
             try:
-                prompt_parts, runtime_model, agent = await _prepare_mem0_branches(
+                prompt_parts, runtime_model, agent = await prepare_mem0_prompt_branches(
                     prepare_memory=lambda: build_memory_prompt_parts(
                         prompt,
                         agent_name,
@@ -1272,8 +1144,8 @@ async def _prepare_agent_and_prompt(
                         execution_identity=execution_identity,
                     ),
                     build_agent=_resolve_model_and_build_agent,
-                    cleanup_agent=lambda built_agent: _close_unreturned_agent(built_agent, scope_context),
                     agent_name=agent_name,
+                    shared_scope_storage=scope_context.storage if scope_context is not None else None,
                     pipeline_timing=pipeline_timing,
                 )
             finally:
