+++ tests/test_event_loop_offloading.py
@@ -200,160 +200,6 @@ async def test_prepare_agent_and_prompt_joins_overlapping_mem0_branches_before_h
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("failing_branch", ["memory", "agent", "both", "memory_cancelled"])
-async def test_prepare_agent_and_prompt_propagates_branch_failure_directly(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    failing_branch: str,
-) -> None:
-    """Either branch failure keeps its original exception and prevents history preparation."""
-    rendezvous = threading.Barrier(2)
-    memory_error = (
-        asyncio.CancelledError("memory cancelled")
-        if failing_branch == "memory_cancelled"
-        else RuntimeError("memory failed")
-    )
-    agent_error = RuntimeError("agent failed")
-    built_agent = MagicMock()
-    built_agent.additional_context = ""
-    prepare_history = AsyncMock()
-    close_unreturned = MagicMock()
-    test_logger = MagicMock()
-
-    async def memory_branch(*_args: object, **_kwargs: object) -> MemoryPromptParts:
-        await asyncio.to_thread(rendezvous.wait, 5.0)
-        if failing_branch in {"memory", "both", "memory_cancelled"}:
-            raise memory_error
-        return MemoryPromptParts()
-
-    def agent_branch(*_args: object, **_kwargs: object) -> MagicMock:
-        rendezvous.wait(5.0)
-        if failing_branch in {"agent", "both"}:
-            raise agent_error
-        return built_agent
-
-    monkeypatch.setattr(ai_module, "build_memory_prompt_parts", memory_branch)
-    monkeypatch.setattr(ai_module, "create_agent", agent_branch)
-    monkeypatch.setattr(ai_module, "prepare_agent_execution_context", prepare_history)
-    monkeypatch.setattr(ai_module, "close_agent_runtime_state_dbs", close_unreturned)
-    monkeypatch.setattr(ai_module, "logger", test_logger)
-
-    config = _prompt_preparation_config()
-    expected_error_type = asyncio.CancelledError if failing_branch == "memory_cancelled" else RuntimeError
-    with pytest.raises(expected_error_type) as raised:
-        await ai_module._prepare_agent_and_prompt(
-            make_turn_context("general"),
-            prompt="hello",
-            runtime_paths=test_runtime_paths(tmp_path),
-            config=config,
-        )
-
-    expected_error = agent_error if failing_branch == "agent" else memory_error
-    assert raised.value is expected_error
-    prepare_history.assert_not_awaited()
-    if failing_branch in {"memory", "memory_cancelled"}:
-        close_unreturned.assert_called_once_with(built_agent, shared_scope_storage=None)
-    else:
-        close_unreturned.assert_not_called()
-    if failing_branch == "both":
-        test_logger.error.assert_called_once()
-    else:
-        test_logger.error.assert_not_called()
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("agent_outcome", ["pending", "finished", "fails", "cancelled"])
-async def test_prepare_agent_and_prompt_cancellation_cleans_agent_build_in_either_order(  # noqa: C901, PLR0915
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    agent_outcome: str,
-) -> None:
-    """Cancellation drains offloaded construction and closes its agent without leaked tasks."""
-    memory_started = asyncio.Event()
-    memory_cleaned = asyncio.Event()
-    agent_started = threading.Event()
-    agent_release = threading.Event()
-    agent_finished = threading.Event()
-    built_agent = MagicMock()
-    built_agent.additional_context = ""
-    close_unreturned = MagicMock()
-    test_logger = MagicMock()
-
-    async def blocked_memory(*_args: object, **_kwargs: object) -> MemoryPromptParts:
-        memory_started.set()
-        try:
-            await asyncio.Event().wait()
-        finally:
-            memory_cleaned.set()
-
-    def blocked_agent(*_args: object, **_kwargs: object) -> MagicMock:
-        agent_started.set()
-        if agent_outcome == "finished":
-            agent_finished.set()
-            return built_agent
-        if not agent_release.wait(5.0):
-            msg = "timed out waiting to release cancelled agent construction"
-            raise TimeoutError(msg)
-        agent_finished.set()
-        if agent_outcome == "fails":
-            msg = "agent failed after cancellation"
-            raise RuntimeError(msg)
-        if agent_outcome == "cancelled":
-            raise asyncio.CancelledError
-        return built_agent
-
-    monkeypatch.setattr(ai_module, "build_memory_prompt_parts", blocked_memory)
-    monkeypatch.setattr(ai_module, "create_agent", blocked_agent)
-    monkeypatch.setattr(ai_module, "prepare_agent_execution_context", AsyncMock())
-    monkeypatch.setattr(ai_module, "close_agent_runtime_state_dbs", close_unreturned)
-    monkeypatch.setattr(ai_module, "logger", test_logger)
-
-    config = _prompt_preparation_config()
-    baseline_tasks = set(asyncio.all_tasks())
-    prepare_task = asyncio.create_task(
-        ai_module._prepare_agent_and_prompt(
-            make_turn_context("general"),
-            prompt="hello",
-            runtime_paths=test_runtime_paths(tmp_path),
-            config=config,
-        ),
-    )
-    try:
-        await asyncio.wait_for(memory_started.wait(), timeout=1.0)
-        assert await asyncio.to_thread(agent_started.wait, 1.0)
-        if agent_outcome == "finished":
-            assert await asyncio.to_thread(agent_finished.wait, 1.0)
-        prepare_task.cancel()
-        await asyncio.wait_for(memory_cleaned.wait(), timeout=1.0)
-        await asyncio.sleep(0)
-        if agent_outcome != "finished":
-            assert not prepare_task.done()
-        if agent_outcome == "pending":
-            prepare_task.cancel()
-            await asyncio.sleep(0)
-            assert not prepare_task.done()
-
-        agent_release.set()
-        with pytest.raises(asyncio.CancelledError):
-            await prepare_task
-    finally:
-        agent_release.set()
-
-    assert agent_finished.is_set()
-    if agent_outcome in {"fails", "cancelled"}:
-        close_unreturned.assert_not_called()
-    else:
-        close_unreturned.assert_called_once_with(built_agent, shared_scope_storage=None)
-    if agent_outcome == "fails":
-        test_logger.error.assert_called_once()
-    else:
-        test_logger.error.assert_not_called()
-    await asyncio.sleep(0)
-    leaked_tasks = {task for task in asyncio.all_tasks() - baseline_tasks if not task.done()}
-    assert leaked_tasks == set()
-
-
-@pytest.mark.asyncio
 async def test_prepare_agent_and_prompt_keeps_file_memory_before_agent_build(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
